src/cloud_function/cf_sentiment_analysis/requirements.txt
functions-framework==3.9.2
google-cloud-language==2.17.2
google-cloud-secret-manager==2.23.3